Move audit creation to a task
