+++ api/edge_api/identities/tasks.py
@@ -1,6 +1,7 @@
 import logging
 import typing
 
+from audit.models import AuditLog
 from environments.models import Environment, Webhook
 from features.models import Feature, FeatureState
 from task_processor.decorators import register_task_handler
@@ -85,3 +86,34 @@ def sync_identity_document_features(identity_uuid: str):
 
     identity.synchronise_features(valid_feature_names)
     EdgeIdentity.dynamo_wrapper.put_item(identity.to_document())
+
+
+@register_task_handler()
+def generate_audit_log_records(
+    environment_api_key: str, identifier: str, user_id: int, changes: dict
+):
+    audit_records = []
+
+    feature_override_changes = changes.get("feature_overrides")
+    if not feature_override_changes:
+        return
+
+    environment = Environment.objects.select_related(
+        "project", "project__organisation"
+    ).get(api_key=environment_api_key)
+
+    for feature_name, change_details in feature_override_changes.items():
+        action = {"+": "created", "-": "removed", "~": "deleted"}.get(
+            change_details.get("change_type")
+        )
+        log = f"Feature override {action} for feature '{feature_name}' and identity '{identifier}'"
+        audit_records.append(
+            AuditLog(
+                project=environment.project,
+                environment=environment,
+                log=log,
+                author_id=user_id,
+            )
+        )
+
+    AuditLog.objects.bulk_create(audit_records)
